# Required fields checking in the parser