… benchmarks/benchmark.cc
… cmake/CMakeLists.txt
… tests/conformance_upb.c
… tests/test_table.cc
… upb/decode.h
… upb/decode_internal.h
… upb/json_encode.c
… upb/msg.c
… upb/msg_internal.h
… upb/msg_test.cc
… upb/msg_test.proto
… upb/table.c
… upb/table_internal.h
… upbc/message_layout.cc
… upbc/message_layout.h
… upbc/protoc-gen-upb.cc