This PR adds required field checking to the parser. While parsing, a check will be performed at the end of each message or sub-message that occurs in the input that all required fields are present. If a special decoder option is set (`kUpb_DecodeOption_CheckRequired`) the parse will fail with a special error code (`kUpb_DecodeStatus_MissingRequired`) if the parse succeeded in every way except for having missing required fields.

This is not a complete required fields solution on its own. In particular:
- This check can throw false positives if a sub-message is parsed as incomplete, but the sub-message re-occurs later in the payload and "fills in" the missing required fields.
- This check can fail to detect missing required fields in a sub-message if we are merging against an existing message. The existing message could have an incomplete sub-message; if the sub-message does not occur in the binary payload, the parser will never visit it and detect that it is incomplete.

For these reasons, we also need a function that will visit the entire message tree checking for required fields (this is coming in a follow-up PR). This separate function will be used as a second pass for the hard cases. Most cases will be able to avoid the second pass -- which is good for performance -- but edge cases such as `MergeFromString()`.

The general implementation strategy here is:
1. The message mini-table stores how many required fields the message has.
2. Required fields always have the lowest hasbits.
3. The parser will form a mask at parse time to check against the hasbits of a message.

There is also a bit of optimization in this PR to avoid perf or code size regression. In particular:
1. For unknown fields we no longer preserve a pointer to the beginning of the field, for the off-chance that the field will be unknown. Instead we walk the pointer backwards to find the beginning of the field when it is unknown (luckily we have just enough info to walk backward through varints).
2. The structure of the parse loop is changed a bit; the condition of the main `while()` loop is clearer. When fasttable parsing is enabled we need to try parsing at least one field before fast dispatch; this case is handled now in a way that doesn't penalize the parser when fasttable is *not* enabled.

Thanks to these optimizations, this PR ends up being code size neutral and perf positive:

```
name                                      old time/op  new time/op  delta
ArenaInitialBlockOneAlloc                 7.00ns ± 1%  6.29ns ± 0%  -10.26%  (p=0.000 n=12+11)
ArenaOneAlloc                             21.7ns ± 1%  20.8ns ± 1%   -3.99%  (p=0.000 n=12+12)
LoadAdsDescriptor_Proto2                  14.7ms ± 2%  15.2ms ± 7%   +3.34%  (p=0.000 n=10+12)
LoadAdsDescriptor_Upb                     3.86ms ± 1%  3.80ms ± 2%   -1.56%  (p=0.001 n=12+10)
LoadDescriptor_Proto2                      247µs ± 3%   247µs ± 1%     ~     (p=0.843 n=12+12)
LoadDescriptor_Upb                        48.6µs ± 3%  47.6µs ± 1%   -2.04%  (p=0.000 n=11+10)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.9µs ± 2%  18.0µs ±18%     ~     (p=0.093 n=10+12)
Parse_Proto2<FileDesc,NoArena,Copy>       31.8µs ±10%  31.3µs ±11%     ~     (p=0.590 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  17.0µs ± 4%  17.8µs ±19%     ~     (p=0.203 n=10+12)
Parse_Proto2<FileDesc,UseArena,Copy>      19.9µs ± 3%  20.0µs ± 4%     ~     (p=0.756 n=10+11)
Parse_Upb_FileDesc<InitBlock,Alias>       9.53µs ± 1%  8.61µs ± 0%   -9.61%  (p=0.000 n=10+10)
Parse_Upb_FileDesc<InitBlock,Copy>        11.4µs ± 0%  10.2µs ± 0%  -10.84%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<UseArena,Alias>        9.93µs ± 1%  9.08µs ± 2%   -8.54%  (p=0.000 n=11+10)
Parse_Upb_FileDesc<UseArena,Copy>         11.7µs ± 1%  10.7µs ± 0%   -9.16%  (p=0.000 n=12+11)
SerializeDescriptor_Proto2                6.49µs ±37%  5.52µs ± 1%  -14.96%  (p=0.001 n=12+11)
SerializeDescriptor_Upb                   11.3µs ± 0%  11.7µs ± 1%   +3.19%  (p=0.000 n=11+11)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
  +0.6%    +176  +0.7%    +176    upb/def.c
    +4.0%    +144  +4.0%    +144    resolve_msgdef
    +1.4%     +32  +1.4%     +32    create_fielddef
  +0.5%     +39  +0.4%     +32    tests/conformance_upb.c
    +0.7%     +38  +0.7%     +38    write_output
    +7.4%      +7  [ = ]       0    stderr
    -9.1%      -6 -23.1%      -6    parse_proto.msg
  +0.2%      +7  [ = ]       0    upb/msg.c
    +4.1%      +7  [ = ]       0    _upb_msg_addunknown
  -0.1%      -6  [ = ]       0    upb/table.c
    [NEW]    +437  [NEW]    +392    upb_strtable_remove2
    -1.7%      -7  [ = ]       0    upb_inttable_sizedinit
    [DEL]    -436  [DEL]    -392    upb_strtable_remove
 -57.1%     -16 -57.1%     -16    [LOAD #2 [RX]]
    +100%      +3  +100%      +3    __libc_csu_fini
   -76.0%     -19 -76.0%     -19    [LOAD #2 [RX]]
  -0.3%     -26  -0.8%     -64    upb/decode.c
    [NEW]    +157  [NEW]    +112    decode_checkrequired
    +3.1%     +16  +3.4%     +16    _upb_decode
    +5.2%     +16  +6.1%     +16    decode_checkenum_slow
    +3.0%     +16  +3.3%     +16    decode_enum_packed
    +3.7%      +9  +8.3%     +16    decode_isdonefallback
   -15.0%     -16 -22.2%     -16    decode_err
    -3.9%    -224  -3.9%    -224    decode_msg
  -5.4%    -134  [ = ]       0    [Unmapped]
  +0.0%     +40  +0.1%    +128    TOTAL
```